Fixed issue with big numbers

+++ Checker.cpp
@@ -23,7 +23,7 @@ Declaration *Checker::look_up(Id id, Block *b)
 
 Type Checker::check_expr(Expr *expr, Block *b)
 {
-    long int num;
+    long long num;
     bool bl;
     if (expr->isNumLiteral(&num))
         return Type::Int;
